Enable full-screen viewing of vispy canvas in Chipstream.
Summary:
Pressing F11 now toggles between fullscreen display of the vispy canvas
in Chipstream. If the fullscreened canvas is selected, Escape closes it.

This implementation may seem wasteful since it makes the application do
twice as much work preparing canvases for viewing, but in reality, it's
great. The other method I tried, updating the parent of the vispy canvas
(this is how Qt does windowing), was hard to get working smoothly since
all the work done on creating a ChipCanvas would be done every time we
toggled our viewing option. This is slow and bad. In contrast, the
implementation in this commit makes application startup a little bit
slower, but after that, we can switch between fullscreen and
not-fullscreen with moar speedz. And we love speedz!!

Also, pressing Escape while the embedded vispy canvas is selected will
no longer close the canvas (which was the vispy canvas's default
response to the Escape key).

Test Plan:
run Chipstream application, spam F11 and Escape and see if fullscreening
behaves as expected.

Reviewers: #willow, cblam

Reviewed By: cblam

Subscribers: cblam

Differential Revision: https://reviews.leaflabs.com/D133