Avoid memory growth by updating fullscreen window only when visible
Summary:
We create a fullscreen window (canvas_popout widget) at init to
accelerate switching between fullscreen and normal views via F11.

We hide the fullscreen window while it is not in use, but prior to
the present commit, we continued to draw new data to canvas_popout,
and that turns out to causes unbounded memory growth behind the
scenes in Qt-land.

The present commit adds a test to check whether canvas_popout is
visible, and avoids calling ChipCanvas.draw_new_data if it is not.

Test Plan:
Launch Chipstream via the Willow GUI, and start playing.  Launch top
in a shell window and monitor system memory consumption.  Focus on
the Chipstream data pane and toggle between normal and fullscreen
views via F11.  Verify that channel data is displayed as would be
anticipated, in both views, and that system memory consumption is
stable in both views.

Reviewers: #willow, sinback, chrono, perry

Reviewed By: #willow, sinback

Differential Revision: https://reviews.leaflabs.com/D185