Sonifier: add a digital gain control slider
Summary:
Enhance the Sonifier stream analysis plugin by adding a slider to control
digital gain.  This feature allows scaling of a low gain signal from a
Willow probe channel in the digital domain.

The default digital gain is unity (1.0), and the maximum value is defined
presently as 10.0, with a step size of 0.1.

Samples are scaled (multiplied) by the selected gain.  If gain is adjusted
such that the signal clips, the value INT16_MIN is output for the negative
swing, and INT16_MAX for the positive swing.

Test Plan:
Build and run the sonifier, specifing the path to the subsampleemu utility
(review pending) as the argument.  Start Sonifying channel 0 to hear a 12-bit
resolution 1 Khz sine wave.  Click on the Digital Gain control and slide it
to the right using either the mouse or keyboard right arrow.  Listen for
increasing volume, and eventually, a change in character as the signal
clips.

Reviewers: #willow, perry

Reviewed By: #willow, perry

Differential Revision: https://reviews.leaflabs.com/D293