UI: modify scrolling behavior
Summary:
Make Chipstream widget zoom isometrically when the mouse wheel is used
with no modifiers, zoom horizontally when the mouse wheel is used with
the Control modifier, and zoom vertically when the mouse wheel is used
with the Shift modifier. Correct label describing scrollwheel behavior
to be consistent with scrolling behavior.

Test Plan: run Chipstream, scroll with / without modifiers

Reviewers: #willow, tasman

Reviewed By: tasman

Subscribers: tasman, #willow

Differential Revision: https://reviews.leaflabs.com/D292