Chipstream: aggregate rapid chip change events
Summary:
When multiple chip change events occur within CHANGE_CHIP_MS
(currently 500) milliseconds, defer sending a
setSubsamples_byChip request to WillowGUI until CHANGE_CHIP_MS
have elapsed.  Otherwise, continue to instantiate the request immediately.

Processing a setSubsamples_byChip request is time-consuming, so we only
want to send them when necessary, and if a Chipstream user is rapidly
clicking on the forward arrow or back arrow, we want to instantiate the
final chip change (soon) after the clicking has stopped.

Depends on D308.

Test Plan:
Run ChipStream and monitor or log stderr.  Click once on the forward
arrow and verify that a setSubsamples_byChip request is sent for the
new chip number.  Rapidly click on the forward arrow, and verify that
setSubsamples_byChip requests aren't sent for every intermediate value
of chip number, but that a request is sent for the final chip number.
Repeat tests for the back arrow.

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Differential Revision: https://reviews.leaflabs.com/D309